♻️ break up the main continuity synchronization function
