+++ packages/atom.io/realtime-server/src/index.ts
@@ -1,9 +1,9 @@
 import type { Store } from "atom.io/internal"
 import type { Json } from "atom.io/json"
 
+export * from "./continuity/prepare-to-sync-realtime-continuity"
 export * from "./ipc-sockets"
 export * from "./realtime-action-receiver"
-export * from "./realtime-continuity-synchronizer"
 export * from "./realtime-family-provider"
 export * from "./realtime-mutable-family-provider"
 export * from "./realtime-mutable-provider"
